+++ plugins/attrchange/ls.attrchange.js
@@ -2,6 +2,8 @@
 	'use strict';
 	if(!window.addEventListener){return;}
 
+	var rAF = window.requestAnimationFrame || setTimeout;
+
 	var addObserver = function(){
 		var connect, disconnect, observer, connected;
 		var lazySizes = window.lazySizes;
@@ -9,6 +11,16 @@
 		var attributes = {'data-bgset': 1, 'data-include': 1, 'data-poster': 1, 'data-bg': 1, 'data-script': 1};
 		var regClassTest = '(\\s|^)(' + lsCfg.loadedClass;
 		var docElem = document.documentElement;
+
+		var setClass = function(target){
+			rAF(function(){
+				lazySizes.rC(target, lsCfg.loadedClass);
+				if(lsCfg.unloadedClass){
+					lazySizes.rC(target, lsCfg.unloadedClass);
+				}
+				lazySizes.aC(target, lsCfg.lazyClass);
+			});
+		};
 
 		var onMutation = function(mutations){
 			var i, len, mutation, target;
@@ -23,11 +35,7 @@
 				}
 
 				if(target && regClassTest.test(target.className)){
-					lazySizes.rC(target, lsCfg.loadedClass);
-					if(lsCfg.unloadedClass){
-						lazySizes.rC(target, lsCfg.unloadedClass);
-					}
-					lazySizes.aC(target, lsCfg.lazyClass);
+					setClass(target);
 				}
 			}
 		};